refactor: make ruleRecord serializable (#1878)
Using `json.[Un]Marchal` is one way of serializing the data and the
easiest one to start with. This change just does the ugly diff renaming
everything to work with those marshalling methods. Do we want to rename
any of these fields while we have to change them anyway?

**What type of PR is this?**
Other

**What package or component does this PR mostly affect?**
all

**What does this PR do? Why is it needed?**
Make the `ruleRecord` serializable to json in preparation for caching
it.