perf: preload .bazelignore before configure, do not use globs (#1886)
The `.bazelignore` should be loaded once upfront and not depend on the
`config.Configurer` invocation order.

`.bazelignore` does not support globs like the gazelle `excludes` so a
simple map lookup instead of glob evaluation is enough and faster.

Do not invoke `config.Configure` on a directory which is
`.bazelignore`d, unlike `excludes` which only applies to sub-directories
within a BUILD.

**What type of PR is this?**
Other

**What package or component does this PR mostly affect?**
all

**What does this PR do? Why is it needed?**

Cleanup