bzlmod/go_work: support relative paths
We have a use case where we have multiple go modules in a single repo,
but adding go.work breaks some existing tooling, if it's added to the
root. We want to add it to another location like: bazel/thirdparty
but then we need to support having entries in go.work that look like:
`../../src/go/rpk`. This patch allows leading parent segments (..) in
go.work files to support our use case.