feat: include package URL in package_info (#1917)
Requires upgrade to the 1.0.0 release of rules_license, which is more
appropriate than a pre-release anyway.

Follow-up to #1852

Note that I'm still unable to gather the resulting data from an end-user
point of view using the aspects provided by rules_license. Will look for
some help on Slack to see if there's still missing plumbing.

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>