Fix patch file creation using Gazelle in diff mode (#1915)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**
  Bug fix


**What package or component does this PR mostly affect?**
  cmd/gazelle

**What does this PR do? Why is it needed?**
This PR adds a workaround to strip trailing newlines from the resulting
diff when running Gazelle in `mode=diff`. This aligns with the expected
behavior of modern diff tools, which will ignore the last newline in a
file. This is implemented to resolve an issue where modifying the end of
a file caused Gazelle to output a diff that was invalid, due to the
newlines. Applying patches generated in this case could not be applied
when using `git apply` and would apply, but with a warning identifying
this issues when using the system `patch` utility.

**Which issues(s) does this PR fix?**

Fixes #1916 

**Other notes for review**
How to Reproduce:
Edit a `BUILD.bazel` file in this repository, and then run `bazel run
//:gazelle -- -mode=diff -patch=temp.patch`. Next, run `git apply -p0
temp.patch`. Without these changes, patch application will fail, due to
the issue described above.