Add gazelle overrides to enable building rclone
I was just toying around with Gazelle and the rclone project [1]. These
changes were necessary, but not sufficient, to get it to build.

I also needed to make a change for hdfs in my MODULE.bazel, but someone
already beat me to it default_gazelle_overrides.bzl. I solved my problem
with `gazelle:proto legacy`, but it looks like the existing override
does something different.

[1]: https://github.com/rclone/rclone/issues/8093