fix(repo): load repositories should fail on duplicates (#1937)
We've recently had issues because we have multiple `go_repository` rules
with the same name, at different versions.

If we run `go mod tidy` and `gazelle update-repos`, the rule that gets
updated isn't always the one that ends up being used in the build.

To solve this, we are planning to add a check in CI to make sure no
duplicate go_repository rules are loaded.

This change is needed in order to use `gazelle/repo` to enable this
feature. This can also be potentially used to help aid others in
migrating to Bzlmod or simply cleaning up their repos.